requirements-base.txt
numpy
python-dotenv
fastapi
uvicorn[standard]
sentence-transformers
PyPDF2
faiss-cpu
groq
python-multipart
